requirements.txt
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
Pygments==2.2.0
pytz==2018.4
